Don't expose .isEmpty()
Summary:
Continuing on the quest of making `ContentState`'s API better, let's remove `isEmpty()`. `ContentState` is an Immutable `Record`, and this function will return true if the record is empty. That is, if `ContentState` has no keys, not if it represents an empty text box.

This is useless of course because `ContentState` should never be "an empty object"— if it ever is we'd find out because everything Draft.js related would pretty much break, but it shouldn't given the methods we provide to build it.

Someone seems to have gotten confused by this and used it to check that if the text field had no text, so I removed that call. This diff will avoid further confusion.

Differential Revision: D25285932

fbshipit-source-id: 976697f32b181032c86fb1141b68e343679be8ac