Paste alt text for images if available, instead of camera emoji
Summary:
Draft pastes images as a camera emoji inline and generates an entity for them. I'm proposing we instead paste the alt text for the image if it's available.

Generally, clients are expected to transform this entity to support images via plugins. This functionality remains untouched since the entity grows to the necessary width

# Why?

This makes pasting emojis from a facebook post into a composer simply work. Alternatives to this involve client side plugins + server side handling to turn these images back into the original emoji. Doable but really feels like the text editor should send the right assets.

More ideally, we would wanna control this behavior via a plugin, but the current architecture makes this nontrivial

# Extra why?

We get a very big win here but this is generally a better experience IMO, for default editors this is a better blank state than a camera icon, which is generally a source of confusion. As a convention, at least on mac, pasting images as plain text will replace the images with the alt text, so there's a precedent for this being the right behavior

Reviewed By: trueadm, voideanvalue

Differential Revision: D25615752

fbshipit-source-id: e970c51d76d64af9c3cd8c60cdccfca4cf9313f7