Don't jump cursor on browser-executed forward word deletion
Summary:
Fixed the issue brought up in this post:

https://fb.workplace.com/groups/1662633360725440/permalink/2312963885692381/?comment_id=2624165677905532

So here's what happened. We usually intercept keyboard commands and handle them ourselves. Specifically, if a user presses `alt+delete`, we intercept that [here](https://www.internalfb.com/intern/diffusion/WWW/browse/master/html/shared/draft-js/component/utils/getDefaultKeyBinding.js?commit=1003255056&lines=99) and execute `keyCommandDeleteWord`.

akanter here has some custom keybindings set up which delete a word forward with `alt+d`. That's not the `delete` key, so we don't intercept this. We could try and intercept it [here](https://www.internalfb.com/intern/diffusion/WWW/browse/master/html/shared/draft-js/component/utils/getDefaultKeyBinding.js?commit=1003255056&lines=68), but `alt+d` usually inserts delta (“δ”) and we don't want to override whatever OS behaviour the user has.

Instead, on browsers that implement InputEvents, we can check if the action the user took was a forward-word-deletion. We detect deletions by diffing the text we see in the DOM after the input, and assume it's a normal deletion, so we move the caret back by the number of characters that were deleted. If it's a forward deletion, we skip this and just not move the caret.

Reviewed By: akanter

Differential Revision: D26282167

fbshipit-source-id: 9d4deaeb9731072affeb5a3ca41b92b8d8a0533a