Scope Safari hack to just Safari users
Summary:
I'm not sure if this hack is even needed any more (it may have been fixed in a newer Safari version), but it's causing issues when displaying read-only rich text (for example using `<XDSRichTextView >`), where the text is indented by 1px. This is causing alignment issues in my app and breaking screenshot tests for a few components after switching them from `CSUIMarkdownRenderer` to `XDSRichTextView`.

This diff updates the CSS file to only apply the Safari hack when the user is actually using Safari, via the `safari` class we automatically add to the body.

See https://fb.workplace.com/groups/xdsquestions/posts/935346250410978

Reviewed By: agoder

Differential Revision: D31066502

fbshipit-source-id: a3c358f401ea5280e8f7a280c3b28e15e955c699